requirements.txt
# Core AI and NLP - Working versions for Python 3.13
langchain==0.3.14
langchain-openai==0.3.11
langchain-community==0.3.14
langchain-text-splitters==0.3.8
langchain-pinecone==0.2.12
langchain-core==0.3.63
langsmith==0.2.11
openai==1.109.1

# Vector Database
pinecone==7.3.0

# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.12
aiofiles==24.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# Data Processing
pandas==2.3.2
numpy==2.3.3
pydantic==2.10.4
email-validator==2.1.1

# HTTP and API
httpx==0.28.1
requests==2.32.3

# Document Processing
pypdf==6.1.1
python-docx==1.1.2
openpyxl==3.1.5
docx2txt==0.8

# Database
sqlalchemy==2.0.36
alembic==1.14.0

# Development
pytest==8.3.4
pytest-asyncio==0.24.0
black==24.10.0
flake8==7.1.1